Move toolbar into shell folder
