+++ core-web/libs/portlets/dot-ema/src/lib/dot-ema-shell/components/edit-ema-navigation-bar/edit-ema-navigation-bar.component.scss
