+++ core-web/libs/portlets/dot-ema/src/lib/dot-ema-shell/components/edit-ema-navigation-bar/edit-ema-navigation-bar.component.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
 import { RouterModule } from '@angular/router';
 
-import { NavigationBarItem } from '../../shared/models';
+import { NavigationBarItem } from '../../../shared/models';
 
 @Component({
     selector: 'dot-edit-ema-navigation-bar',
